+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Version 2022-dev
 -  switch to using unique_ptr for cgmoleculedef (#652)
 -  switch boundary conditions to unique_ptr (#654)
 -  switch NBList to unique_ptr (#659)
+-  switch csgapplication worker to unique_ptr (#660)
 -  let unique_ptrs manage bead memory on heap (#656)
 
 Version 2021-rc.2 (released XX.01.21)